Adding functionality for collecting feedbacks using Like/Dislike button and field for entering more detailed feedbacks
