+++ src/gradio/app/main.py
@@ -1,16 +1,18 @@
-import asyncio
-import logging
 import uuid
 
-import pymupdf
 import uvicorn
-from fastapi import FastAPI, HTTPException, status
-from langchain_community.llms import Ollama
+from fastapi import FastAPI
 
-import gradio as gr
-
-from .settings import settings
-from .summarized_text import summarize_text
+from .utils import (
+    detailed_feedback,
+    generate_post_async,
+    gr,
+    logging,
+    process_pdf_and_summarize,
+    reset_summarization,
+    simple_feedback,
+    stop_llm,
+)
 
 llm_tasks = {}
 summary_tasks = {}
@@ -18,187 +20,6 @@ summary_tasks = {}
 logging.basicConfig(level=logging.INFO)
 
 app = FastAPI()
-
-
-def validate_pdf_file(file_content):
-    if not file_content:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail="No file was uploaded."
-        )
-    if not getattr(file_content, "name", "").lower().endswith(".pdf"):
-        raise HTTPException(
-            status_code=status.HTTP_415_UNSUPPORTED_MEDIA_TYPE,
-            detail="Uploaded file is not a PDF.",
-        )
-
-
-def extract_text_from_pdf(file_content):
-    try:
-        pdf_document = pymupdf.open(file_content, filetype="pdf")
-        text = ""
-        for page_number in range(pdf_document.page_count):
-            page = pdf_document.load_page(page_number)
-            text += page.get_text("text")
-        pdf_document.close()
-        return text
-    except Exception as e:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Error processing the PDF file: {str(e)}",
-        )
-
-
-async def process_pdf_and_summarize(file_content, session_id) -> str:
-    global summary_tasks
-    logging.info(f"Starting PDF processing for session {session_id}...")
-
-    validate_pdf_file(file_content)  # validation input
-
-    text = extract_text_from_pdf(file_content)  # extract text from pdf
-
-    try:
-        loop = asyncio.get_running_loop()
-        task = loop.run_in_executor(None, summarize_text, text)
-        summary_tasks[session_id] = task
-        output_text = await task
-        logging.info("Summarization completed.")
-        return output_text
-    except Exception as e:
-        logging.error(f"Failed during processing: {str(e)}")
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail=f"Failed to process PDF: {str(e)}",
-        )
-
-
-def prepare_post_text(
-    summary,
-    audience,
-    english_level,
-    length,
-    hashtag_preference,
-    perspective,
-    emoji_usage,
-    question_option,
-    paper_url,
-    custom_requirements,
-):
-    post_text = (
-        "Can you create a LinkedIn post, summarize the text: "
-        + summary
-        + "; In the post use these parameters:"
-    )
-
-    if english_level:
-        post_text += f" Use {english_level} English."
-    if length:
-        post_text += f" Length of post: {length}."
-    if emoji_usage:
-        post_text += f" {emoji_usage}."
-    if audience:
-        post_text += f" Targeted at {audience}."
-    if hashtag_preference:
-        post_text += f" {hashtag_preference}."
-    if perspective:
-        post_text += f" Written from a {perspective} perspective."
-    if question_option:
-        post_text += f" {question_option}"
-    if paper_url:
-        post_text += f" URL to original paper: {paper_url}"
-    if custom_requirements:
-        post_text += f" Custom requirements: {custom_requirements}"
-
-    post_text += "\n Skip introduction about audience, do not greet audience, finish with call to action"
-
-    return post_text
-
-
-async def generate_post_async(
-    summary,
-    audience,
-    english_level,
-    length,
-    hashtag_preference,
-    perspective,
-    emoji_usage,
-    question_option,
-    paper_url,
-    custom_requirements,
-    session_id,
-):
-    global llm_tasks
-
-    if not summary:
-        raise HTTPException(
-            status_code=status.HTTP_403_FORBIDDEN,
-            detail="Summarization not completed",
-        )
-    else:
-        post_text = prepare_post_text(
-            summary,
-            audience,
-            english_level,
-            length,
-            hashtag_preference,
-            perspective,
-            emoji_usage,
-            question_option,
-            paper_url,
-            custom_requirements,
-        )
-
-        llm = Ollama(
-            model=settings.generation_ollama_model,
-            base_url=settings.ollama_url,
-        )
-
-        try:
-            loop = asyncio.get_running_loop()
-            llm_task = loop.run_in_executor(None, llm.invoke, post_text)
-            llm_tasks[session_id] = llm_task
-            generated_post = await llm_task
-            return generated_post
-        except Exception as e:
-            logging.error(f"Failed post generation process: {str(e)}.")
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail="Failed post generation process.",
-            )
-
-
-def stop_llm(session_id):
-    global llm_tasks
-    task = llm_tasks.get(session_id)
-
-    if task is not None:
-        llm_tasks.pop(session_id, None)
-        logging.info(
-            f"The post generation process for session {session_id} was cancelled."
-        )
-    else:
-        logging.info(
-            f"No post generation process was found running for session {session_id}."
-        )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail="No post generation running"
-        )
-
-
-async def reset_summarization(session_id):
-    global summary_tasks
-    task = summary_tasks.get(session_id)
-
-    if task is not None:
-        summary_tasks.pop(session_id, None)
-        logging.info(f"Summarization task for session {session_id} was cancelled.")
-    else:
-        logging.info(
-            f"No summarization process was found running for session {session_id}."
-        )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST, detail="No summarization running"
-        )
-
 
 with gr.Blocks(title="Research dissemination assistant") as demo:
     session_id = gr.State(lambda: uuid.uuid4().hex)
@@ -211,8 +32,18 @@ with gr.Blocks(title="Research dissemination assistant") as demo:
         </div>
         """
     )
+
+    gr.HTML(
+        """
+        <p style="color: #4a4a4a;">
+            Generate a social media post quickly from your research paper with the Quick Research Post tool.
+        </p>
+    """
+    )
+
     first_part = gr.Group()
     second_part = gr.Group()
+    feeback_part = gr.Group()
 
     with first_part:
         gr.Markdown("## Document Summarization")
@@ -274,6 +105,38 @@ with gr.Blocks(title="Research dissemination assistant") as demo:
         submit_button = gr.Button("Generate post")
         stop_button = gr.Button("Stop")
         post_output = gr.Markdown(label="Generated LinkedIn Post")
+
+    with feeback_part:
+        gr.Markdown("## Feedback")
+        with gr.Row():
+            like_button = gr.Button("👍")
+            dislike_button = gr.Button("👎")
+
+        feedback_text = gr.Textbox(
+            visible=False,
+            label="Write your detailed feedback here:",
+            lines=4,
+            interactive=True,
+        )
+        submit_feedback_button = gr.Button(
+            "Submit Feedback", visible=False, interactive=True
+        )
+
+    like_button.click(
+        fn=lambda session_id: simple_feedback("like", session_id),
+        inputs=[session_id],
+        outputs=[feedback_text, submit_feedback_button],
+    )
+
+    dislike_button.click(
+        fn=lambda session_id: simple_feedback("dislike", session_id),
+        inputs=[session_id],
+        outputs=[feedback_text, submit_feedback_button],
+    )
+
+    submit_feedback_button.click(
+        fn=detailed_feedback, inputs=[feedback_text, session_id], outputs=[]
+    )
 
     click_event = start_summarization_button.click(
         fn=process_pdf_and_summarize,
